Uniform Confidence Bands for a Marginal Treatment Effect (MTE) Function .

unif_gen computes uniform confidence bands of the MTE function given the parametric part of MTE is known. For details, see Tsuda, Jin and Okui (2024+).

Usage

unif_gen(
  outcome_name = outcome_name,
  covariate_name = covariate_name,
  instrument_name = instrument_name,
  treatment_name = treatment_name,
  P_hut_name = P_hut_name,
  beta0 = beta0,
  beta1 = beta1,
  data = data,
  bw = bw,
  l_eval = l_eval,
  u_eval = u_eval,
  covariate_value = covariate_value,
  significance_level = 0.05
)

Arguments

outcome_name

The name of the outcome

covariate_name

The name of covariate

instrument_name

The name of instrument variables

treatment_name

The name of a treatment variable

P_hut_name

The name of estimated propensity scores.

beta0

A vector of parameters of a linear model for the outcome given an individual is not treated

beta1

A vector of parameters of a linear model for the outcome given an individual is treated

data

data.frame including propensity scores to be used for the estimation of MTE

bw

The bandwidth used in estimating the nonparametric part of MTE

l_eval

The minimum value of evaluation points for the estimated MTE

u_eval

The maximum value of evaluation points for the estimated MTE

covariate_value

A covariate value to plot the estimate MTE

significance_level

The significance level for the construction of uniform confidence band. The default is 0.05.

Value

A list that contains the following elements:

Estimation

A data frame of the following results:
evaluation point of MTE, estimated value of MTE, standard error of MTE, Lower bound of a confidence band , Upper bound of a confidence band

critical_value

A critical value corresponding to the siginificance level chosen

Plot

A list that contains the ggplot elements for unifrom confidence band

References

Tsuda, Jin and Okui (2024+). Uniform Confidence Band for Marginal Treatment Effect Function. will be available on arXiv

Examples

if (FALSE) {
set.seed(1)
library(MASS)
SS<-2000
X1<-rnorm(n=SS,mean=0,sd=1)
X2<-rnorm(n=SS,mean=0,sd=1)
X3<-rnorm(n=SS,mean=0,sd=1)
X4<-rnorm(n=SS,mean=0,sd=1)
Z1<-rnorm(n=SS,mean=0,sd=1)
Z2<-rnorm(n=SS,mean=0,sd=1)
Z3<-rnorm(n=SS,mean=0,sd=1)
Z4<-rnorm(n=SS,mean=0,sd=1)
E0<-c(1,0.5,0.3)
E1<-c(0.5,1,0.8)
E2<-c(0.3,0.8,1)
E<-cbind(E0,E1,E2)
E4<-mvrnorm(n=SS,mu=c(0,0,0),Sigma=E)
U0<-E4[,1]
U1<-E4[,2]
V<-E4[,3]
Expl<-cbind(X1,X2,X3,X4)
Instr<-cbind(X1,X2,X3,X4,Z1,Z2,Z3,Z4)
PInstr<-cbind(Z1,Z2,Z3,Z4)
beta0<-c(0.5,0.1,-0.1,-0.5)
beta1<-c(0.8,0.4,-0.4,-0.8)
ganma0<-c(0.4,0.4,0.4,0.4,0.3,0.3,0.3,0.3)
D<-numeric(SS)
for(i in 1:SS){
  if ((ganma0 %*% (Instr[i,]))>V[i]){
    D[i]=1}
}
Y<-numeric(SS)
for(i in 1:SS){
  Y[i]<-D[i]*((beta1%*%Expl[i,])+U1[i])+(1-D[i])*((beta0%*%Expl[i,])+U0[i])
}
demo=as.data.frame(cbind(Y,Expl,PInstr,D))
Supp_P_hut_U<-0.85
Supp_P_hut_L<-0.15
O_SS<-1000
B2<-seq(Supp_P_hut_L,Supp_P_hut_U,length=O_SS)
covariate_value<-c(0,0,0,0)

RESULT_P_hut<-p_hut_gen(outcome_name="Y",
                        covariate_name=c("X1","X2","X3","X4"),
                        instrument_name=c("Z1","Z2","Z3","Z4"),
                        treatment_name= "D",
                        data=demo,
                        family="probit",
                        trim=0.01,
                        intercept=TRUE
)

RESULT_unif<-unif_gen(outcome_name="Y",
                      covariate_name=c("X1","X2","X3","X4"),
                      instrument_name=c("Z1","Z2","Z3","Z4"),
                      treatment_name= "D",
                      P_hut_name= "P_hut",
                      bw=0.1,
                      beta0=beta0,
                      beta1=beta1,
                      data=RESULT_P_hut$Data,
                      l_eval=0.15,
                      u_eval=0.85,
                      covariate_value=covariate_value,
                      significance_level=0.05
)
}